DRA-1946 move extractValuesFromPreservicaContent to ExtractedPreservicaValues
