+++ src/main/java/dk/kb/present/util/DataCleanup.java
@@ -14,21 +14,12 @@
  */
 package dk.kb.present.util;
 
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import dk.kb.present.util.saxhandlers.ElementsExtractionHandler;
 import dk.kb.util.DatetimeParser;
 import dk.kb.util.MalformedIOException;
-import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.xml.sax.SAXException;
 
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 import java.time.ZonedDateTime;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -71,25 +62,4 @@ public class DataCleanup {
         }
     }
 
-    /**
-     * Extract all needed values from a preservica record. These values are either tricky values such as dates, where we know that extra parsing is needed or values that are
-     * used in multiple parts of the processing of the record.
-     * @param content of the record. i.e. the XML data.
-     * @param recordId of the processed record. Used for logging and debugging.
-     * @return a {@link ExtractedPreservicaValues}-object containing the extracted values.
-     */
-    public static ExtractedPreservicaValues extractValuesFromPreservicaContent(String content, String recordId) throws ParserConfigurationException, SAXException {
-        try (InputStream xml = IOUtils.toInputStream(content, StandardCharsets.UTF_8)) {
-            SAXParserFactory factory = SAXParserFactory.newInstance();
-            factory.setNamespaceAware(false);
-            SAXParser saxParser = factory.newSAXParser();
-
-            ElementsExtractionHandler handler = new ElementsExtractionHandler(recordId);
-            saxParser.parse(xml, handler);
-
-            return handler.getDataValues();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
 }
